Allow more generic usage of chart widget
This commit does not restrict the chart passed to the `ChartWidget` to a `&mut` reference to something that implements `Chart`. This commit instead allows the `ChartWidget` to be generic over any type that implements `Chart` and also adds a blanket implementation over all `&mut T where T: Chart`. These two changes in conjunction allow for the `ChartWidget` to take either a mutable reference to or ownership of a type that implements `Chart`.